Move completion-lay-hilit fix out of org. Pos-frame which-key
The fix as it was would only actually load after org was loaded, and there is no
reason for that. Moving it out ensures it gets executed even without org-mode.

I like how the pos-frame which-key looks, so enabled.

Also we incorporate some of the changes for the spacemacs file from the template